refactor: extract form styles into form.module.css
